review: clean up after moving
- remove uneeded headers
- rename to new directory
